Work with TODO and fix warning on optimize process

+++ components/Optimizer.php
@@ -141,7 +141,7 @@ class Optimizer extends \JustCoded\WP\ImageOptimizer\core\Component {
 	 * Ajax function for manual image optimize
 	 */
 	public function manual_optimize() {
-		$attach_id = (int)$_POST['attach_id'];
+		$attach_id = (int) $_POST['attach_id'];
 
 		$this->optimize_images( [ $attach_id ] );
 		$model           = new Media();
@@ -164,10 +164,10 @@ class Optimizer extends \JustCoded\WP\ImageOptimizer\core\Component {
 	 */
 	protected function optimize_images( array $attach_ids ) {
 		global $wp_filesystem;
-		$media      = new Media();
-		$log        = new Log();
+		$media               = new Media();
+		$log                 = new Log();
 		$before_attach_found = $attach_ids;
-		$attach_ids = $media->size_limit( $attach_ids );
+		$attach_ids          = $media->size_limit( $attach_ids );
 		// add filter for WP_FIlesystem permission.
 		add_filter( 'filesystem_method', array( $this, 'filesystem_direct' ) );
 		WP_Filesystem();
@@ -183,37 +183,41 @@ class Optimizer extends \JustCoded\WP\ImageOptimizer\core\Component {
 		// upload images from service.
 		$dir = WP_CONTENT_DIR . '/tmp/';
 		\JustImageOptimizer::$service->upload_optimize_images( $attach_ids, $dir );
-
-		$get_image = scandir( $dir );
-		$get_path  = $media->get_uploads_path();
-		// process image replacement.
-		if ( ! empty( $get_image ) ) {
-			foreach ( $get_image as $key => $file ) {
-				if ( $wp_filesystem->is_file( $dir . $file ) ) {
-					foreach ( $get_path as $path ) {
-						if ( $wp_filesystem->exists( $path . '/' . $file ) ) {
-							$optimize_image_size = getimagesize( $dir . $file );
-							if ( 25 < $optimize_image_size[0] && 25 < $optimize_image_size[1] ) {
-								$wp_filesystem->copy( $dir . $file, $path . '/' . $file, true );
-								$log->save_status( $request_id, $file, Log::STATUS_OPTIMIZED );
-							} else {
-								$log->save_status( $request_id, $file, Log::STATUS_ABORTED );
+		//check if folder not empty
+		if ( count( glob( $dir ) ) != 0 ) {
+			$get_image = scandir( $dir );
+			$get_path  = $media->get_uploads_path();
+			// process image replacement.
+			if ( ! empty( $get_image ) ) {
+				foreach ( $get_image as $key => $file ) {
+					if ( $wp_filesystem->is_file( $dir . $file ) ) {
+						foreach ( $get_path as $path ) {
+							if ( $wp_filesystem->exists( $path . '/' . $file ) ) {
+								$optimize_image_size = getimagesize( $dir . $file );
+								if ( 25 < $optimize_image_size[0] && 25 < $optimize_image_size[1] ) {
+									$wp_filesystem->copy( $dir . $file, $path . '/' . $file, true );
+									$log->save_status( $request_id, $file, Log::STATUS_OPTIMIZED );
+								} else {
+									$log->save_status( $request_id, $file, Log::STATUS_ABORTED );
+								}
 							}
 						}
 					}
 				}
+				self::delete_dir( $dir );
+				// set statistics and status after replace images.
+				foreach ( $attach_ids as $attach_id ) {
+					$file_sizes = $media->get_file_sizes( $attach_id, 'detailed' );
+					$media->update_stats( $attach_id, $file_sizes );
+					$log->update_details( $request_id, $attach_id, $file_sizes );
+					update_post_meta( $attach_id, '_just_img_opt_status', Media::STATUS_PROCESSED );
+				}
 			}
-			self::delete_dir( $dir );
-		}
-
-		// TODO: if images processing failed - clean status to IN QUEUE again, otherwise stats is wrong (bytes_after = 0 and it calculate wrong stats)
-
-		// set statistics and status after replace images.
-		foreach ( $attach_ids as $attach_id ) {
-			$file_sizes = $media->get_file_sizes( $attach_id, 'detailed' );
-			$media->update_stats( $attach_id, $file_sizes );
-			$log->update_details( $request_id, $attach_id, $file_sizes );
-			update_post_meta( $attach_id, '_just_img_opt_status', Media::STATUS_PROCESSED );
+		} else {
+			foreach ( $attach_ids as $attach_id ) {
+				$log->update_status( $attach_id, $request_id, Log::STATUS_REMOVED );
+				$media->clean_statistics( $attach_id );
+			}
 		}
 	}
 }
